pci: fix libpciaccess vendor/device ID for vir...
pci: fix libpciaccess vendor/device ID for virtual function devices in case of SR-IOV on Linux

libpciaccess shows 0xffff for vendor/device IDs for virtual functions.
Use Linux sysfs files to work-around the issue when available.

The actual fix requires to look for SR-IOV info in the physical
function, which requires root access on Linux. Left as TODO.

This commit was SVN r5627.