pci: fallback to linuxpci when libpciaccess fai...
pci: fallback to linuxpci when libpciaccess fails to initialize.

RHEL5 and FC9-12 have a buggy libpciaccess that only works for root.
Detect this early in pci instantiate() so that we can ignore the pci
component entirely and avoid uselessly conflicting with the linuxpci
one.

No need to do that for pciutils since no such bug is known
(if pciutils fails, linuxpci will likely fail as well), and
because splitting the pciutils init between instantiate() and
discover() would be harder because of the setjmp exit-workaround.

This commit was SVN r5631.