rename: Fix the renaming of hwloc_fls/ffs*
hwloc_flsl/ffsl may be macros or static inlines depending on the platform.
We can't easily rename this because renaming occurs before the definition,
we don't know yet if it will be a macro or not, and we don't rename macros
(or we would redefine the #define).

Add an intermediate name in the case of static inline
(hwloc_ffsl_from_ffs32, hwloc_flsl_manual, etc.).
Rename those always-static-inline name.
And #define the final hwloc_flsl/ffsl to those, so that
hwloc_ffsl/flsl are always #define, and don't need renaming.

This prevents hwloc_ffsl/flsl symbols to miss the renaming on some
platforms.

This commit was SVN r5643.