pci: Always use a 256-byte config space
Initialize everything to 0xff (unknown) before reading the config space,
remove useless size checks everything, and rely on the callee to actually
check things it reads (it already did).

libpci doesn't report the length of the config space it actually reads,
so we can't safely rely on config space size checks. So this actually stop
using unitialized config space data when not running as root
(between 65th and 256th bytes, where we look for capabilities and link speed).

This commit was SVN r5747.