headers: move the header file documentation to ...
headers: move the header file documentation to groups

We don't have a file index, so putting doxygen docs in the file
header is almost useless. Move/add interesting things to the
doxy groups underneath so that it actually appears in the generated
doc.

By the way, update the README.

This commit was SVN r5792.