configure: Fix the disabling of doxygen when building from tarballs
We check if we are in a developer tree, display the result, but do nothing with it.
Add the relevant code to update enable_doxygen accordingly if it wasn't forced
on the command-line.

This was unnoticed for a long time because we distribute hwloc.tag,
which causes tarball-builds not to regenerate the doc.
But they will regenerate it as soon as a dependency changes.

Now doxygen is entirely disabled properly in tarball-builds for real
as long as --enable-doxygen is not given.