Linux: allow to read uname from fake fs root
Read a new "fake" /proc/hwloc-nofile-info where we can store things
that cannot be saved as a file. Put the output of uname there when
gathering the topology and use it on input to overwrite local or
missing data.

This improves testing under tests/linux by allowing architecture
specific code to run from fake fsroot. Will be useful for advanced
cpuinfo parsing.

We manually remove /proc/hwloc-nofile-info in test-gather-topology.sh
because it only partially brings fake fsroot information.
Removing it creates the same !is_thissystem() output from the fake
fsroot and from the real machine.

Once we gather other information in /proc/hwloc-nofile-info:
* sysconf(_SC_LARGE_PAGESIZE);
* hwloc_getpagesize();
* hwloc_fallback_nbprocessors(topology);
we may be able to really use /proc/hwloc-nofile-info in test-gather-topology.sh
and really discover the exact same thing from the saved topology and
from the original machine. We'll see that later.