Linux: use the uname architecture to better match cpumodel information
Look at the uname architecture to select which cpuinfo parsing function
will be used.
For now, we have one for x86/x86_64/mic, one for ia64 and one for arm.
Others (or when the architecture is unknown, when fsroot was changed
and there's no /proc/hwloc-nofile-info) fallback to the default "dumb"
parsing function.