hwloc_distrib(): don't ignore objects when they are too many of them
When we distrib N processes among M objects with M>N,
don't ignore the last M-N objects.

Rather scatter the ignored objects all along the array.
If one chunk is rounded-up, the immediate next one may be rounded-down to 0.
If so, it doesn't get ignored anymore, its cpuset is merged with the cpuset
of the previous one before being returned.

Thanks to Tim Creech for suggesting that this well known issue
must be fixed.