Add --disable-cpuid configure flag to work around buggy processor simulators
Some simulators report invalid CPUID information, causing crashes in
the x86 code. It's not clear how many issues we would have to work around
in the code, so we don't try to solve them dynamically.

By the way another workaround is setting HWLOC_COMPONENTS=-x86 in the environment.

Thanks for Andrew Friedley for reporting the issue.
In his case, Simics seems to report that no CPU is attached to some local cache.