Only show local memory in non-verbose attributes
Multiple levels of memories are coming, so displaying the total memory
in each level will be confusing.
Change attr_snprintf() to only show the local memory of the current
object in non-verbose mode (verbose mode already shows local and total).

Then update lstopo to add the total memory when total memory > local memory.
So lstopo is back to the same state except that it's now clear that
the displayd memory sizes are local except when "total" is added.