doxy: integrate hwloc.doxy and netloc.doxy through a main.doxy
Add a global introduction, and the common install instructions
and strip things from hwloc and netloc sections.
Add references in all directions to ease things.

One thing that remains to be updated/merged is the bug report section,
we'll fix this later once the actual merge is done and once we know
where netloc bug reports should be sent.