configure: improve _HWLOC_CHECK_DECL()
We want to check whether a function declaration is already present
in one of the header. So we redeclared the function with an absurd
prototype. Unfortunately, clang 3.2 and icc only warn without failing
in this case.

This commit redeclares the function as a variable instead of a function
with a different prototype, and this leads to the expected failure with
GNU, clang and icc.

strncasecmp() is now properly detected in many more cases.

As long as the _HWLOC_CHECK_DECL is invoked on a function and not on a
global variable, it should be fine.

See http://www.open-mpi.org/community/lists/devel/2014/09/15775.php
and next messages in the thread

Thanks to Gilles Gouaillardet for the patch.