Add netloc
Library source files go under netloc/* on the side of hwloc/
Headers, tests, utils are added near hwloc ones.

doc/main.doxy is added, which includes the former hwloc.doxy and the new netloc.doxy
Will be better inntegrated in next commit so that we don't touch hwloc too much here.

Netloc is enabled by default when supported (basically everywhere except
on Windows and Solaris), can be disabled with --disable-netloc.
It builds a separate libnetloc shared library (so that we can stabilize the
netloc ABI without breaking the hwloc ABI, at least for now).
If you disable netloc or do not include/link with netloc, you get a "normal" hwloc.

netloc test files are not included because some of them are big.
They may go to a separate repository (we may want to do that for
hwloc's tests/linux/ at some point if 1.7MB is considered too much).

hwloc/master is now branch v2.0