core: insert_object_by_parent() doesn't need to check children order, only XML does
Most use cases of insert_object_by_parent() are just inserting I/O and Misc,
no cpuset ordering there.
Out-of-order when duplicating would require a fix in the source topology,
not at duplicate/insert time.
XML is the only case when out-of-order insert_by_parent() can really occur,
move the warning to the XML code, and improve the error message.