insert: traverse the children list only once during insert
Now that the children list is guaranteed in order and without intersection,
directly do the insert in a single loop across the children list.

We only insert the new object as a new child at the very end, so that we don't have
to remove it on error later.

The only possible error is an intersection between the new object and an
existing child, and it may happen after we already moved some child below
the new object. We move those new children back.