pci: uniformize the insertion of the group for hostbridge
Don't try to look for the exact parent. It would be annoying
since some {complete_,}{cpu,node}sets don't exist yet.

The insert code already deals with those sets correctly.
So just insert a group and let the core merge it if needed.

That's what we already do for inserting distance-groups.