Add a GroupType info key for Groups
Set to Book for Linux/s390 books.
Displayed as "Book" in lstopo graphics and "Group(Book)" in lstopo textual.

We only replace Group with this attribute value in lstopo but not in the core
(otherwise we'd have to handle it back in hwloc_obj_type_sscanf() etc).

(cherry picked from commit 2ed31c9b0cc2645476690846f629e6fabdd0e420)

Conflicts:
	NEWS
	include/hwloc.h
	tests/hwloc/linux/20s390-2g6s4c.output