core: when users insert a Group after load(), apply ignore rules immediately
We can't use the normal code path for ignoring after load()
because we want to know if that precise Group was dropped or not,
and that would be overkill anyway.

But after load(), we can apply ignore rules immediately
since nothing else can be added later to change the
keep_structure value of an object.
So just ignore immediately during insert if needed.

By the way, clarify that Groups are ignored keep_structure by default.