bitmap: optimize comparison functions for same-size bitmaps
isequal(), intersects(), isincluded(), compare_first() and compare()

Optimize for the common case where bitmaps have same size.
Then handle different-sizes in a special case.
Thus we don't check whether a ulong is allocated in a bitmap everywhere.

compare_first() is used heavily in the core, it can be responsible
for a large part of the discovery CPU usage.