core: also merge group keep_structure during insert during load()
Commit 03776b6322ce2237c9bc029eb98e325b099e0984 introduce this
immediate merge after load() but there's actually no reason
to prevent it before load() since the structure cannot change later
if this group is identical to another object.

This prevents useless reconnecting of levels because Groups were added.
It also simplify other code path by not having Groups in the middle of
the topology (for instance the upcoming PCI locality quirk).