pci-numa: quirk for buggy BIOS on dual Xeon E5v3 with culster-on-die
It looks like many BIOSes export buggy PCI-NUMA locality when cluster-on-die
is enabled on dual Xeon E5v3. Each package has 2 NUMA nodes, but only the first
NUMA node has some PCI links. These BIOS confuse the 1st NUMA node of 2nd package
with 2nd NUMA node of 1st package.

Detect this case, and warn, and fix it.
The user may disable the quirk by forcing the locality of this PCI bus
as explained in the warning.

Walking up the tree for finding the NUMA node and package numbers is easy
because commit 5995e9649fdba8ed220bce8463428693399af8d2 removed spurious
Group in the middle of the topology.