utils: fix races between install-man and install-exec-hook
install-exec-hook runs in another instance of make,
so duplicate dependencies between install-man and install-exec-hook
are not resolved, causing failures in parallel builds.

utils/hwloc doesn't need the install-exec-hook->install-man dependency anymore,
remove it.

utils/lstopo needs it, so move the man-related exec-hook to install-data-hook,
which runs at the end of install-data, which depends on install-man first.

Thanks to Peter Korsgaard for the report and patch against utils/hwloc.