tests/xml: add a test with irregular grouping
Restricted SGI UV made of blades with 2 sockets (creating a group)
and blades with one socket and some I/Os (no group),
with some offline CPUs.

Force C locale in the console output so that we can check the distance
matrix (it could be truncated if the core doesn't handle the group/NUMA
levels correctly).