XML: make sure XML callbacks are initialized everywhere
Just call components_init/fini() when we're not directly
operating on a valid topology.

The topology argument passed to xmldiff load/export/destroy
is actually unused now. For export, it might be good to
remove it entirely to simplify things (if you don't know
the reference topology in advance, you need to load the diff
to find out which topology to load, but you needed a topology
to load the diff). For others, it doesn't seem to matter.
Unless we find a possible reason to keep it, I guess we could
remove it before 2.0.