pci: ignore bridges with invalid secondary-subordinate buses
Some buggy PCI bridges report invalid secondary-subordinate in their
config space, for instance 00-00, which means they would contain
the master bus.

Primary bus number bugs are not uncommon, we can workaround them by
overwritting with what's in the bus ID.
Secondary-subordinate bus bugs cannot be fixed, and they raise
asserts in the insert algorithms because of cycles.

Add some basic checks for all these bus numbers and ignore the bridge
entirely if failed.
Ideally we would also check that [secondary-subordinate] is included
in parent [secondary+1:subordinate] but that's harder.

Thanks to George Bosilca for reporting the issue at
  http://www.open-mpi.org/community/lists/devel/2015/09/17993.php
on a machine with a
  Pericom Semiconductor PCI Express to PCI-XPI7C9X130 PCI-X Bridge
whose bus ID is 22:00.0 while primary/secondary/subordinate buses are 00
according to the config space.