bind.c: fix silly compiler warning
clang on OS X warns that p may be uninitialized.  Best I can figure,
it's assuming posix_memalign() may fail to fill &p and still return an
errno of 0.  Technically -- on a buggy system -- that can happen.  So
we might as well initialize p to be NULL just to cover all cases / be
defensive.