API: replace IO topology flags with set_type_filter()
WHOLE_IO is KEEP_ALL on Bridge, PCI, OS and Misc.

IO_DEVICES and IO_BRIDGES use a new variant of filtering; KEEP_IMPORTANT
For PCI/OS, it keeps "likely-important" devices.
For Bridges, it keeps bridges with children.

The overall behavior (default and in lstopo) is unchanged
but it's now more flexible and uniform.
One noticeable change is that hostbridges are not kept
anymore when bridges are totally ignored. There's no point
in enforcing the hierarchy ->hostbridge->...bridges...->pci->os
anymore since we can now remove any of these types individually.