pci: replace HWLOC_PCI_<domain>_<bus>_LOCALCPUS env var with HWLOC_PCI_LOCALITY
The former only works for devices matching the given bus, or if we have
all intermediate bridges up to a matching bus.
The new one support ranges that just work even if without bridges,
accept multiple localities at once, and can be stored in a file.

HWLOC_PCI_<domain>_<bus>_LOCALCPUS is still accepted when HWLOC_PCI_LOCALITY
isn't set, but it shows a warning. Otherwise it is ignored.