x86: mark max_log_proc as legacy and restrict its use
It has been deprecated on AMD for a while, and it may even be lower
than MaxNumberofCores when cpuid 0x80000008 is supported.

On Intel, x2APIC gives better information, and it's limited
to 256 threads per package anyway.

Make it clear that max_log_proc should only be used for old feature detection.