x86: Fix L3 sharing on 12-core AMD Bulldozer (Opteron 62x) and Piledriver (63xx)
These newer processors have the same APIC id problem as Magny-Cours.
We need a quirk to compute cacheids properly because a L3 may start on an
APIC id that is not a multiple of the number of ids in the processor.

This is a workaround to many "intersection between L3 and Package" errors
reported on AMD 62xx and 63xx in the past. It may finally be fixed in the
Linux kernel in the near future.

Thanks to Fabian Wein, Joseph Mingrone and Sergio Manzetti, and others
for providing cpuid dumps and testing the code.

(cherry picked from commit 8d1e418e54ea4b4d99bd4de557d3db952571228b)