core: allow Groups to be replace with identical objects
Directly replace an existing group object when an identical
object is inserted.
We already supported the reverse case: ignoring the insertion
of a group when an existing object is identical.

Useful for the x86 because it creates groups for AMD compute units
that are later replaced with L2 objects.
It didn't appear in the output because groups were later removed
as their level does not bring any structure.
We now remove them earlier.