linux: ignore the KNL L3 reported by sysfs
The size is always 0, so it's not useful. And shared_cpu_map
is always the entire machine, which is wrong for SNC2/SNC4 modes.

Moreover this L3 wouldn't be inserted at the right place, it would
end up as the same place as the package. But we want the package to be
above both DDR and MCDRAM nodes, while we would want this L3 to appear
below the DDR node. So don't bother quirking the package and L3 vertical
positions, just ignore the L3.

The actual information about that memory-side cache is provided by
hwloc-dump-hwdata, and we already insert it exactly where it goes by
specifying both cpuset and nodeset.