x86: add caches that don't exist yet
Useful for the Solaris backend on x86 since it doesn't detect any caches.
Thanks to Ryan Zezeski for reporting the issue.

We didn't do it in the past to avoid conflicts between OS and x86 backends
reporting different objects of same types. The new code check whether the current
PU already has a cache before inserting a new one here. In theory we could still
create a conflict if the newly-added cache spans on another PU (with higher os_index)
that already has a cache of same type. If such strange topology bugs ever
appear, we'll tell user to disable the x86 backend while looking for a better fix.

Fixes #144.