linux: don't require libnuma anymore
Many users reported failures to get membind support because libnuma devel
headers are not installed. We only need libnuma because that's where
numa syscalls are wired, let's hardwire them to remove the dependency.

We still use libnuma for the specific interoperability testing in
linux-libnuma.h, but not in the core anymore.

The main drawback is that we now always use the syscall() macro
(previously only used when libc didn't define sched_setaffinity),
that means we require a working syscall with 6 arguments (for mbind).
It looks like it has been working for a long time (libnuma doesn't
use it before 2.11 but the relevant x86_64 issue was fixed in glibc
2.3.4 in 2004). At least, it works with 2.7 from Debian Lenny
(early 2009).

Another drawback is that we assume all these functions and flags
are available. Those were added to the kernel in 2006, so it's fine too.
If using an older kernel, support will say that things are supported
but we may actually return ENOSYS.

By the way, fix the way we check syscall() in configure
(syscallX needs X+1 arguments).