linux: optimize AMD compute unit coreid vs thread_siblings quirk
AMD dual-core compute units report 2 thread siblings per core,
but those have different core IDs, so we need a special quirk.
Fortunately, this only occurs on AMD Family 15h processors,
and those won't be mixed with other families in the same machine.

So only check for that case in the first core and remember the
result for all other cores in the machine.
This avoids duplicate reading of core_id topology files on all
hyperthreaded machines.