distances: complete rework of the API
Both query and add/remove functions are now in hwloc/distances.h

Objects can disappear or be reordered between distance addition
and distance queries by user (because of ignoring, restrict, ...).
We don't want to always reupdate things whenever the core removes objects.
This made v1.x core very complex, while distances are not actually
widely used.

For v2.x, rework the logic to simplify the core:
Objects are actually needed only for grouping and for user queries.
Grouping is now performed when the distances is added, because objects
are valid at that point of time.
If the user ever queries distances information, we'll lookup objects
again in case they changed.
In the meantime, we use the new gp_index (or the os_index when
guaranteed unique) to identify objects.

The new API manipulates objects instead of indexes for similar reasons.
That means adding distances by user may only be done after load().
Not an issue since this operation is rather a topology change
than a topology configuration.

New distances are attached to the topolopy and link whatever set of
same-type objects (instead of being attached to a "root" object and
covering *all* its children).
We can even have multiple distance matrices for objects of the same
type if ever needed (the API would even support objects of different types).

Distances may have a kind (given by OS or user)|(latency or bandwidth)
that is informative and used for grouping.

Adding a distance also lets you enable grouping or not,
accurately or not.

Distance addition through environment variable is not possible anymore.
It will be done with hwloc-annotate now.

The restrict flag ADAPT_DISTANCES is removed.
Restricting is difficult now, and we only do it on user query anyway.

We don't normalize anymore.

The main drawback is that we cannot import all v1.x distances from XML.
But we can still important case: NUMA distances covering the entire machine.

However, v1.x cannot import new 2.x distances from XML.
We'll have to either export specifically for v1.x,
or export both v2 and v1.x in the XML.