xml: fix empty userdata contents import when not decoding
Userdata with length==0 are exported without any content at all,
but the hwloc core doesn't want the imported buffer to be NULL.
Don't let the get_content() callback leave the buffer uninitialized
when there's no content. Rather set it to "", which will be reexported
back to an empty content.

This bug caused a segfault in strdup(NULL) in hwloc_utils_userdata_import_cb()
when not decoding XML userdata in a -O0 build.
Observed with nolibxml (on 16-2gr2gr2n2c+misc.xml) but could break
with libxml too since it didn't initialize the contents buffer either.

Only matters for hwloc tools that import and reexport XML userdata
without decoding/reencoding it (lstopo and hwloc-annotate).

Thanks to Samuel Thibault.