tests: add common infrastructure for running non-distributed tests
... and enable it under tests/hwloc/linux/extra

We'll be able to have tests in the extra/ subdirectory that are
ignored in make dist, and in make check from tarballs.

We manually call the LOG_DRIVER automake command-line. LOG_DRIVER is
officialy documented and can be superseded by a custom script.
So it's unlikely its current command-line options will change in
the near future.

Extra check outputs are similar to the official automake outputs,
except for color because I didn't prefix the LOG_DRIVER command-line
with $(am__tty_colors). Don't know if that variable may change in the future.