lstopo/draw: recurse twice in the tree, but only once in each node
Instead of having each node recurse once for preparing
and once for drawing, have them recurse only once.
But the top-level recursion is initiated twice, once
for preparing and once for drawing.

No need to temporarily change the drawing mode in each node anymore
(introduced in previous commit).

Thanks to Ahmad Boissetri Binzagr for the draft.