lstopo/draw: remove obsolete checks for custom children without cpuset
Custom is gone, normal children always have a cpuset,
no need to invoke the network drawing code in such a case.

By the way, drop support for drawing groups of machines
as network since the system of groups isn't drawn as network.
Document that we'll want to properly detect the general case
 "system of groups (of groups ...) of machines"
before readding that code.