API: add obj->gp_index
This is a new index that is guaranteed unique (across the entire topology)
and persistent during topology changes (restrict, grouping) and dump/reload.

The logical_index is unique but may change during restrict/grouping/...
The os_index doesn't change but it doesn't exist for most levels,
and it's not unique across the entire topology (x86 core ids are only unique
within a given package).

This new index is mostly for internal use but could still be used by application
if really needed.

We have to filter it out from XML outputs under tests/hwloc/linux/ because
gp_index depends on the discovery order, which can vary from one pciaccess
release to another.