API+linux: add new THISSYSTEM_ALLOWED_RESOURCES flag
XML is useful for improving discovery time, but it doesn't work
when the XML should be adapted to different cgroups for different
jobs on the same machine.

This new flag (combined with IS_THISSYSTEM) (or the corresponding env var)
load normal XML or synthetic and apply cpu/cgroup restrictions that the
actual machine reports on Linux.

Note that the XML must be created without any restriction (outside of any
cgroup that disallows some PUs or NUMAs on Linux).

Signed-off-by: Brice Goglin <Brice.Goglin@inria.fr>