linux: cleanup the parsing of cpu topology indexes
Use open instead of fopen, the former is much faster on such sysfs files (about 25% on KNL),
the latter doesn't ease anything when reading one line of ten chars.

Also return an int when sysfs files are filled with %d.
Don't force to -1 in case of failure to parse the file, just let the caller use the right default value.

Signed-off-by: Brice Goglin <Brice.Goglin@inria.fr>