core: add TMA and use it for allocating objects during dup()
TMA is an optional "topology-specific memory allocator".
It will be used for duplicating topologies to shared-memory segment.

That only actual shmem allocator will take care of freeing things
during destroy() by destroying the entire shmem region.
So there's no need to convert frees and reallocs to this optional TMA,
and that would have required to convert all allocations in the core,
including those outside of the dup() case.

So, only allocations inside dup() are converted.
That allocator will also never fail (because we precompute the required shmem length).
so error cases for failed allocation do not need to be converted to that allocator.

Signed-off-by: Brice Goglin <Brice.Goglin@inria.fr>