utils: fix the searching of CPU-less objects
We fixed the search for CPU-less NUMA nodes in hwloc-calc, hwloc-bind and friends
in e6b9e7f3116beae5c0c44a96accccb8680418d6a (in 1.11.3) but we broke it again in
3df864a1c5519719c5efa47d528464359e7213a7 (in 1.11.5) because issue #215 required
CPU-less objects to be ignored in hwloc_get_*obj*_inside_cpuset_by_*. Indeed
hwloc-calc and friends use these functions and therefore cannot find CPU-less
NUMA nodes anymore.

Fix this by looking for both cpuset and nodeset in utils instead of only cpuset.

In theory, an object cannot be CPU-less and node-less, but in practice it can
because of cgroups. Such objects still need to be ignored.

Thanks to Gilles Gouaillardet for the report.

Fixes #233.

Signed-off-by: Brice Goglin <Brice.Goglin@inria.fr>