linux: fix the ignoring of the sysfs L3 on KNL
Early KNLs expose a L3 in sysfs using cpuid, but its attributes are wrong
(size is always 0 and sharing is the full machine) because they are not
updated depending on Cluster/Memory mode.

We ignore that L3 and use hwloc-dump-hwdata instead, but that
ignoring was broken in 03ea6c6360f91da7d069dfbe196f65c64550f507.

Fortunately, this doesn't seem needed anymore on current KNLs.

Signed-off-by: Brice Goglin <Brice.Goglin@inria.fr>