API: add hwloc/shmem.h for duplicating a topology to shmem and adopting from shmem
It requires all processes to mmap() the shmem region at the exact same virtual address
(because of all pointers we have inside hwloc objects).
The caller takes care of finding that address before using this API,
for instance by looking in /proc/<pid>/maps of all processes.

In practice, this API is designed to reduce the memory footprint on machines
with limited physical memory. Given the huge size of virtual address spaces
on current platforms, most of the virtual space is empty when the physical
memory is small. And the layout of address spaces of multiple processes on a
same machine is likely similar. So finding a common hole isn't hard in the
vast majority of cases.

OpenMPI uses heuristics such as the middle of largest hole in the first process
(about 50TB on KNL) and assuming it will be available in other processes too,
which seems to work fine so far.
See orte/mca/rtc/hwloc/rtc_hwloc.c in OMPI git master for an example.

Thanks to Ralph Castain for co-designing this API and implementing the
OpenMPI port to test it.

Signed-off-by: Brice Goglin <Brice.Goglin@inria.fr>