solaris: handle lgrp_latency_cookie() errors
It can return -1 when some NUMA nodes are not available to the current process,
ignore the distance matrix entirely is that case.
We could also reduce the matrix to the available nodes instead of entirely dropping it.

By the way, handle calloc() failures for the matrix allocation.

Signed-off-by: Brice Goglin <Brice.Goglin@inria.fr>

(cherry picked from commit a4c15e2131898faa94cf19d425174c1182a32c64)