lstopo --ps: show only comm of processes
With module systems, spack, etc. the path to application binaries get longer
and longer, and eventually the actual name of the binary goes away on the
right in cmdline. Conversely, many options might be passed on the
cmdline. Finding the actual binary name is in the end at best a
heuristic.

It is thus safer to just show the comm of the process (15 characters).

This reworks cmdline/comm/stat parsing, to use /proc/pid/comm when it is
available (since 14 Dec 2009), or /proc/pid/stat otherwise.