core: move NUMA nodes to memory children list
This is finally the big memory rework that was required KNL broke things.

We could support other kind of "memory" objects such as memory-side caches
but we don't have any such object type yet.
Therefore the memory child hierarchy is just a single NUMA node on most
machines, or two on KNL.

In most machines, NUMA nodes are attached to the machine or package objects.
In dual-NUMA packages (Intel CoD, AMD Magny-Cours/Bulldozer, etc),
it's attached to L3.
When there's no normal parent with the exactly required cpuset,
we must insert a group (or kind "memory").
If groups are filtered-out, we have to attach to a larger parent,
which means the NUMA locality can become too wide,
or we can also get two local NUMA nodes if ignoring too many things.

Memory objects are only attached to normal objects, not to I/O or Misc.
I/O are only attached to normal objects or I/O, not to Memory or Misc.
Misc can be attached anywhere.

Backends may insert NUMA nodes by cpuset as usual, and the core will take
care of adding a group (if needed) and attaching the NUMA node below.
Or they may use a new hwloc__attach_memory_object() to attach to a
specific parent (e.g. on KNL).

Memory objects have cpuset that matches where they are attached.
Their nodesets match the NUMA node(s) they include (i.e. only themselves for now).

CPU-less NUMA nodes have to be attached a CPU-less group below root.
They cannot be attached to the root because it would look like they
are local to all PUs.

Corner cases to fix:
* If NUMA is attached to a PU for some reason (single core xeon, restrict, etc),
  and if removing that PU during restrict, we cannot remove the PU because it
  has a NUMA child.
  We should put a CPU-less group instead of the PU and move the NUMA node there.
* KNL dual-NUMA configs aren't properly exported to v1 XML yet,
  the MCDRAM is ignored for now.
* Backends should be reworked to insert CPU-side objects first, then memory,
  just like we do for I/Os.
* Synthetic import still support NUMA, and inserts groups if needed.
  Synthetic export currently ignores memory children.
* lstopo just displays memory children as any children for now, which means
  NUMA nodes are listed after other CPU-side objects.

Signed-off-by: Brice Goglin <Brice.Goglin@inria.fr>