distances: don't warn if distances root couldn't be found because cpuset is empty
If the agregate cpuset of distance objects is empty, we cannot find a root to
store that distance.
This may happen if all objects are CPU-less and/or if some objects were removed
because of KEEP_STRUCTURE filtering.

Thanks to Olivier Aumage for the reporting on issue
Was on a dual-P8+ machine with only one PU in the cgroup,
ignoring all types with KEEP_STRUCTURE.

Signed-off-by: Brice Goglin <Brice.Goglin@inria.fr>