tests: disable shmem test on OSX
10.11+ disables "System Integrity Protection" by default
(and it's not easy to disable in our CI VMs).
It disables DYLD_LIBRARY_PATH inheriting, which causes this
test to fail because the child process fails to find libhwloc.so
because it doesn't get libtool env vars.

Could be fixed by passing the full exec path to system()
instead of $argv[0].

Signed-off-by: Brice Goglin <Brice.Goglin@inria.fr>