plugins: add a enabled_by_default flag
Some components may want to be installed in the default location
but only be enabled explicitly, either because they add some debugging
info, or because they add information to the topology that only
specific users want/need.

Add a enable_by_default field in the component which is set in our
components but may be unset in those other components.

Signed-off-by: Brice Goglin <Brice.Goglin@inria.fr>