lstopo/draw: add --children-order=plain to disable displaying of memory children above
By default, we display NUMA nodes above other children so that
the relative locality of PU and NUMA appears vertically just like
in hwloc 1.x.

With this new option, we can also display all children together.
It means NUMA nodes and CPU-side objects are listed below
their common parents. That's more precise, but less readable for
basic users on basic topologies.

This option may be extended later to specify whether all children
are displayed together, or split in categories in a specific order.
Would be useful for avoiding strange rectangular placements when
there's 1 L3 and 3 I/O (with different sizes) below a package,
for instance.

We call this "order" to distinguish from "layout" which is used
to specific vertical/horizontal/rectangular placement of siblings.

Signed-off-by: Brice Goglin <Brice.Goglin@inria.fr>