core: os_index is unsigned, update alloc_setup_object() accordingly
Use the new HWLOC_UNKNOWN_INDEX instead of -1,
and document it as the default when os_index is unknown or irrelevant.

Add casts everywhere needed for unsigned os_index.

We can't put HWLOC_UNKNOWN_INDEX in an enum since it would be implemented
as a signed int, requiring cast everywhere. But it doesn't really need
documentation anyway, so keep it as a constant macro.

Signed-off-by: Brice Goglin <Brice.Goglin@inria.fr>