xml/export/v1: export non-first memory children too
At least for KNL, we must be able to export non-first local NUMA children
to v1, which means it must be placed on the side with empty cpusets.

In theory, we should also clear non-first nodeset bits from normal
children nodeset (the MCDRAM isn't considered local in v1),
but that's not strictly necessary since v1.x will fix those bits during import

If we have this v2 tree
     |
  objectX
     |-----|---------|------------|
  childY childZ memorychildR memorychildQ
we export to v1 as
     |----------------|
  memorychildR   memorychildQ
     |
   objectX
     |-----|
  childY childZ
so that NUMA node end up *above* caches (and packages)
since those are their parent in most cases in v2.

When objectX has siblings, we also insert a Group above all this
to englobe memorychildR and Q for KNL clusters.
Otherwise we get out-of-order insert warnings in v1 import
because MCDRAM of cluster#0 (cpuless) is inserted before its sibling
DDR of cluser#1 (not cpuless).

Signed-off-by: Brice Goglin <Brice.Goglin@inria.fr>