xml/import/v1: don't add an intermediate Group above KNL MCDRAM
We add intermediate Group above NUMA node when their locality
is different from the parent (e.g. tests/linux/128ia64-17n4s2c
and tests/xml/8ia64-2n2s2c+1n) but that's not the case on KNL
since v1.x couldn't set the desired locality.

Detect that specific MCDRAM and properly attach it near DDR
as expected with v2 on KNL.

Ideally we would also detect similar cases on future non-KNL
platforms with multiple local NUMA nodes.
That's unlikely to occur with v1.x.
And we have no way to be sure if this CPU-less node is desired or not.

Signed-off-by: Brice Goglin <Brice.Goglin@inria.fr>