bitmap: return int (and errors) from all functions that may fail
In case realloc'ing the array of ulongs ever fails.
free(), zero() and fill() are the only functions that still return void.

Even singlify() can fail if ever called on a bitmap that contains
0 in all preallocated ulongs and the infinite set
(we need to alloc an additional ulong and set to 0x1).

from_ulong() doesn't currently ever fail but it could if we
ever preallocate 0 ulongs.

Clsoes #205.

Signed-off-by: Brice Goglin <Brice.Goglin@inria.fr>