API: add hwloc_topology_abi_check()
If two parts of a single process ever want to use the same topology
structure in memory, and if they don't use the same hwloc library,
the topology structure layout may be different.

Use-case would be OpenMP using a MPI-provided topology which was
generated by hwloc embedded in OpenMPI, while OpenMP uses another
hwloc.

If ever called on a v1.x topology, this function won't crash
since it reads ->topology_abi at the very beginning of the
topology structure. We had the number of levels there in the
past, we won't read anything similar to the expected 0x20000
topology ABI version.

Signed-off-by: Brice Goglin <Brice.Goglin@inria.fr>