tests: add hwloc_topology_abi
Check hwloc_topology_abi_check()

And check that the layout (offsetof/sizeof) of some fields in
struct hwloc_topology (as well as bitmaps) has not changed on
Linux/x86_64, so that we detect when HWLOC_TOPOLOGY_ABI must
be bumped.

This 2nd part can be disabled with HWLOC_IGNORE_TOPOLOGY_ABI if
some compiler/flags ever require to relax those ABI checks.

Signed-off-by: Brice Goglin <Brice.Goglin@inria.fr>