distances: update some docs
get() is not thread-safe in the general case because it
may have to refresh distances if the topology was modified
(and refresh is not thread-safe).

We could add a public refresh() call that multithreaded
apps should call after modifying the topo in a thread-safe env
and before calling get() in a parallel env.

But it can be emulated with a (dummy but valid) get().

Signed-off-by: Brice Goglin <Brice.Goglin@inria.fr>